optimize verify flow
